requirements.txt
fastapi
uvicorn
google-generativeai>=0.3.0
selenium>=4.10.0
webdriver-manager>=4.0.0
requests>=2.31.0
aiohttp>=3.8.5
python-dotenv>=1.0.0
asyncio>=3.4.3
regex>=2023.6.3